191023 first

+++ 03_django/03_django_form/articles/views.py
@@ -1,11 +1,12 @@
+import hashlib
 from IPython import embed
+from django.contrib.auth import get_user_model
 from django.shortcuts import render, redirect, get_object_or_404, get_list_or_404
 from django.contrib.auth.decorators import login_required
 from django.http import Http404, HttpResponse
 from django.views.decorators.http import require_POST
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
-
 # Create your views here.
 
 
@@ -39,9 +40,12 @@ def create(request):
         # form 인스턴스를 생성하고 요청에 의한 데이터를 인자로 받는다. (binding)
         # 이 처리과정은 binding 이라고 불리며 유효성 체크를 할 수 있도록 해준다.
         form = ArticleForm(request.POST)
+        # embed()
         # form.is_valid() 로 form 이 유효한지 체크한다.
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             # form.cleaned_data 로 정제된 데이터를 받는다.
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
@@ -64,11 +68,12 @@ def create(request):
 def detail(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     comments = article.comment_set.all()  # article 의 모든 댓글
+    person = get_object_or_404(get_user_model(), pk=article.user_id)
     comment_form = CommentForm()  # 댓글 폼
     # article = Article.objects.get(pk=article_pk)
 
     context = {'article': article,
-               'comment_form': comment_form, 'comments': comments, }
+               'comment_form': comment_form, 'comments': comments, 'person': person, }
     return render(request, 'articles/detail.html', context)
 
 
@@ -76,26 +81,32 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect(article)
     return redirect('articles:index')
 
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            # article.title = form.cleaned_data.get('title')
-            # article.content = form.cleaned_data.get('content')
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                # article.title = form.cleaned_data.get('title')
+                # article.content = form.cleaned_data.get('content')
 
-            article.save()
-            return redirect(article)
+                article.save()
+                return redirect(article)
+        else:
+            # articleForm 을 초기화 (이전에 DB에 저장된 )
+            form = ArticleForm(instance=article)
+        # 1. POST : 검증에 실패한 form(오류 메세지도 포함된 상태)
+        # 2. GET : 초기화된 form
     else:
-        # articleForm 을 초기화 (이전에 DB에 저장된 )
-        form = ArticleForm(instance=article)
-    # 1. POST : 검증에 실패한 form(오류 메세지도 포함된 상태)
-    # 2. GET : 초기화된 form
+        return redirect('articles:index')
     context = {'form': form, 'article': article, }
     return render(request, 'articles/form.html', context)
 
@@ -109,6 +120,7 @@ def comments_create(request, article_pk):
             # 객체를 Create 하지만, db 에 레코드는 작성하지 않는다.
             comment = comment_form.save(commit=False)
             comment.article_id = article_pk
+            comment.user = request.user
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -117,6 +129,41 @@ def comments_create(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if request.user == comment.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)
     return HttpResponse('FIRE!!!!!', status=401)
+
+
+@login_required
+def like(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+
+    # get 을 쓰면 없을 때 오류가 난다. 그래서 filter 를 쓴다.
+    if article.like_users.filter(pk=request.user.pk).exists():
+        article.like_users.remove(request.user)
+    else:
+        article.like_users.add(request.user)
+
+        #  해당 게시글에 좋아요를 누른 사람들 중에서 현재 접속유저가 있다면 좋아요를 취소
+    # if request.user in article.like_users.all():
+    #     article.like_users.remove(request.user)  # 좋아요 취소
+    # else:
+    #     article.like_users.add(request.user)    # 좋아요
+    return redirect('articles:index')
+
+
+@login_required
+def follow(request, article_pk, user_pk):
+    # 게시글 유저
+    person = get_object_or_404(get_user_model(), pk=user_pk)
+    # 접속 유저
+    user = request.user
+    # 내(request.user)가 게시글 유저 팔로워 목록에 이미 존재한다면,
+    # if user in person.followers.all()
+    if person != user:
+        if person.followers.filter(pk=user.pk).exists():
+            person.followers.remove(user)
+        else:
+            person.followers.add(user)
+    return redirect('articles:detail', article_pk)
